backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
beautifulsoup4==4.12.3
requests==2.31.0
langchain==0.2.0
langchain-google-genai==1.0.5
python-dotenv==1.0.1
pydantic==2.7.1